Add files via upload
This is the new version of the first code file which I used to generate streamlines. The issue appears to be with the "connectivity_matrix" function, there's apparently a syntax error in the definition of the function that causes this error. I'm not sure why this would be an issue because as far as I'm aware, the definition of the function was not changed in the DIPY library from when I used it a couple months ago.